backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Data Science (Python 3.13 compatible)
pandas
numpy>=2.1.0
openpyxl==3.1.2
scikit-learn>=1.5.0
pyarrow>=18.0.0

# Forecasting
prophet==1.1.5

# Others
supabase==2.4.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic>=2.8.0
pydantic-settings==2.2.1

# Image & Maps
Pillow>=10.3.0
googlemaps>=4.10.0

# Optional (comment out if causing issues)
# sentence-transformers>=3.0.0
# torch>=2.4.0